src/cats_queue/requirements.txt
kafka-python
faiss-cpu
tqdm
openvino